requirements.txt
flask
Flask-SQLAlchemy
psycopg2
sqlalchemy-stubs
types-Flask
typing_extensions
